Added more tests to increase overall test coverage.
Limit the number of threads to avoid starving the runner

Apply Python coverage

Move Python to the self-hosted machine

Fixed venv typo

Specify the Python version

Fixed the line that was misreporting bad Python data

Fixed bad coverage command

This is the real fix this time

Create README.md

Fixed the first set of issues

Fixed other remaining issues

Fixed build issues from build

Removed const_casts

Added tests for the data utilities

Added tests for the test of the UnrealDataLoader

Added a test for parsing the schema

Added test for the importing libraries

Added missing files

+++ Source/RPGMenus/RPGMenus/Private/Primatives/DisplayText.cpp
@@ -29,7 +29,7 @@ void UDisplayText::SetTextInfo() {
 
 		if (SizeBox != nullptr) {
 			auto TextPadding = GetDisplayTextPadding();
-			SizeBox->SetHeightOverride(GetTextSize().Y + TextPadding.Top + TextPadding.Bottom);
+			SizeBox->SetHeightOverride(static_cast<float>(GetTextSize().Y) + TextPadding.Top + TextPadding.Bottom);
 		}
 	}
 }
@@ -46,7 +46,7 @@ void UDisplayText::SetText(const FText& NewText) {
 
 		if (SizeBox != nullptr) {
 			auto TextPadding = GetDisplayTextPadding();
-			SizeBox->SetHeightOverride(GetTextSize().Y + TextPadding.Top + TextPadding.Bottom);
+			SizeBox->SetHeightOverride(static_cast<float>(GetTextSize().Y) + TextPadding.Top + TextPadding.Bottom);
 		}
 
 		OnTextSet(NewText);
